+++ socialhome/content/tests/test_models.py
@@ -99,6 +99,7 @@ class TestContentModel(SocialhomeTestCase):
                 "parent": "",
                 "profile_id": "",
                 "content_type": self.public_content.content_type.string_value,
+                "has_shared": False,
             },
             {
                 "id": self.site_content.id,
@@ -127,8 +128,15 @@ class TestContentModel(SocialhomeTestCase):
                 "parent": "",
                 "profile_id": "",
                 "content_type": self.site_content.content_type.string_value,
+                "has_shared": False,
             }
         ])
+
+    def test_has_shared(self):
+        self.assertFalse(Content.has_shared(self.public_content.id, self.local_user.profile.id))
+        # Do a share
+        self.public_content.share(self.local_user.profile)
+        self.assertTrue(Content.has_shared(self.public_content.id, self.local_user.profile.id))
 
     def test_is_local(self):
         self.assertFalse(self.site_content.is_local)
@@ -164,6 +172,16 @@ class TestContentModel(SocialhomeTestCase):
         self.assertEqual(share.share_of, self.public_content)
         self.assertNotEqual(share.id, self.public_content.id)
         self.assertEqual(share.author, self.local_user.profile)
+
+    def test_unshare_raises_if_no_share_exists(self):
+        with self.assertRaises(ValidationError):
+            self.public_content.unshare(self.profile)
+
+    def test_unshare_removes_a_share(self):
+        self.public_content.share(self.local_user.profile)
+        assert Content.has_shared(self.public_content.id, self.local_user.profile.id)
+        self.public_content.unshare(self.local_user.profile)
+        self.assertFalse(Content.has_shared(self.public_content.id, self.local_user.id))
 
     def test_fix_local_uploads(self):
         self.public_content.text = "foobar ![](/media/uploads/12345.jpg) barfoo"
@@ -228,6 +246,7 @@ class TestContentModel(SocialhomeTestCase):
             "parent": "",
             "profile_id": "",
             "content_type": self.public_content.content_type.string_value,
+            "has_shared": False,
         })
 
         # Add a share
@@ -265,6 +284,7 @@ class TestContentModel(SocialhomeTestCase):
             "parent": "",
             "profile_id": self.public_content.author.id,
             "content_type": self.public_content.content_type.string_value,
+            "has_shared": False,
         })
 
     def test_dict_for_view_edited_post(self):
@@ -297,6 +317,7 @@ class TestContentModel(SocialhomeTestCase):
                 "parent": "",
                 "profile_id": self.public_content.author.id,
                 "content_type": self.public_content.content_type.string_value,
+                "has_shared": False,
             })
 
     def test_short_text(self):